Group and move things around in the sign module.
