+++ src/sign/crypto/openssl.rs
@@ -12,6 +12,7 @@
 #![cfg_attr(docsrs, doc(cfg(feature = "openssl")))]
 
 use core::fmt;
+
 use std::{boxed::Box, vec::Vec};
 
 use openssl::{
@@ -23,9 +24,9 @@ use openssl::{
 use secrecy::ExposeSecret;
 
 use crate::base::iana::SecAlg;
-use crate::sign::{
-    GenerateParams, RsaSecretKeyBytes, SecretKeyBytes, SignError, SignRaw,
-};
+use crate::sign::error::SignError;
+use crate::sign::keys::keypair::GenerateParams;
+use crate::sign::{RsaSecretKeyBytes, SecretKeyBytes, SignRaw};
 use crate::validate::{PublicKeyBytes, RsaPublicKeyBytes, Signature};
 
 //----------- KeyPair --------------------------------------------------------
@@ -449,11 +450,12 @@ mod tests {
 
     use crate::{
         base::iana::SecAlg,
-        sign::{GenerateParams, SecretKeyBytes, SignRaw},
+        sign::{SecretKeyBytes, SignRaw},
         validate::Key,
     };
 
     use super::KeyPair;
+    use crate::sign::keys::keypair::GenerateParams;
 
     const KEYS: &[(SecAlg, u16)] = &[
         (SecAlg::RSASHA256, 60616),
