+++ src/sign/keys/mod.rs
@@ -2,3 +2,4 @@ pub mod bytes;
 pub mod keymeta;
 pub mod keypair;
 pub mod keyset;
+pub mod signingkey;
